rename thing-if-sdk to thing-if
