+++ browser-test/ThingIFAPIWithoutDependecies.html
@@ -5,7 +5,7 @@
         <link rel="stylesheet" href="https://code.jquery.com/qunit/qunit-2.0.1.css">
         <script src="https://code.jquery.com/qunit/qunit-2.0.1.js"></script>
         <script src="lib/promise.js"></script>
-        <script src="lib/thing-if-sdk.js"></script>
+        <script src="lib/thing-if.js"></script>
         <script src="js/TestApp.js"></script>
         <script src="js/ThingIFAPIWithoutDependencies.js"></script>
     </head>
