+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "thing-if-sdk",
+  "name": "thing-if",
   "version": "1.0.0",
   "description": "Javascript SDK for Kii Thing Interaction Framework.",
   "main": "lib/ThingIFSDK.js",
