+++ test/large/APIAuthorOnboardTest.ts
@@ -6,7 +6,7 @@ import {testApp} from './utils/TestApp';
 import { TestInfo } from './utils/TestInfo';
 
 declare var require: any
-let thingIFSDK = require('../../../dist/thing-if-sdk.js');
+let thingIFSDK = require('../../../dist/thing-if.js');
 
 describe("Large Tests for APIAuthor#Onboarding:", function () {
