+++ test/large/ThingIFAPIQueryTest.ts
@@ -11,7 +11,7 @@ import { NotEqualsClauseInQuery, AllClause } from '../../src/QueryClause';
 import { RangeClauseInTrigger } from '../../src/TriggerClause';
 
 declare var require: any
-let thingIFSDK = require('../../../dist/thing-if-sdk.js');
+let thingIFSDK = require('../../../dist/thing-if.js');
 thingIFSDK.Logger.getInstance().setLogLevel("debug");
 
 describe("Large Tests for ThingIFAPI#query:", function () {
