+++ test/large/ThingIFAPIThingStateTest.ts
@@ -7,7 +7,7 @@ import { ThingIFAPI } from '../../src/ThingIFAPI'
 import { TestInfo } from './utils/TestInfo';
 
 declare var require: any
-let thingIFSDK = require('../../../dist/thing-if-sdk.js');
+let thingIFSDK = require('../../../dist/thing-if.js');
 
 describe("Large Tests for get/update thingType of ThingIFAPI:", function () {
