+++ test/large/ThingIFAPIThingTest.ts
@@ -6,7 +6,7 @@ import {testApp} from './utils/TestApp';
 import {ThingIFAPI} from '../../src/ThingIFAPI'
 
 declare var require: any
-let thingIFSDK = require('../../../dist/thing-if-sdk.js');
+let thingIFSDK = require('../../../dist/thing-if.js');
 
 describe("Large Tests for thing APIs(ThingIFAPI):", function () {
